misc: lint

+++ contracts/interfaces/IWETH.sol
@@ -2,28 +2,23 @@
 pragma solidity ^0.7.6;
 
 interface IWETH {
-    function deposit() external payable;
+  function deposit() external payable;
 
-    function withdraw(uint256) external;
+  function withdraw(uint) external;
 
-    function balanceOf(address account) external view returns (uint256);
+  function balanceOf(address account) external view returns (uint);
 
-    function transfer(address recipient, uint256 amount)
-        external
-        returns (bool);
+  function transfer(address recipient, uint amount) external returns (bool);
 
-    function allowance(address owner, address spender)
-        external
-        view
-        returns (uint256);
+  function allowance(address owner, address spender) external view returns (uint);
 
-    function approve(address spender, uint256 amount) external returns (bool);
+  function approve(address spender, uint amount) external returns (bool);
 
-    function transferFrom(
-        address sender,
-        address recipient,
-        uint256 amount
-    ) external returns (bool);
+  function transferFrom(
+    address sender,
+    address recipient,
+    uint amount
+  ) external returns (bool);
 
-    function decimals() external view returns (uint256);
+  function decimals() external view returns (uint);
 }
